+++ src/main/java/com/akiba/backend/market/dto/response/MarketPostSimilarResponse.java
@@ -1,10 +1,14 @@
 package com.akiba.backend.market.dto.response;
 
 import lombok.Builder;
+import java.math.BigDecimal;
 import lombok.Getter;
+import java.math.BigDecimal;
 
 import java.time.LocalDateTime;
+import java.math.BigDecimal;
 import java.util.List;
+import java.math.BigDecimal;
 
 @Getter
 @Builder
@@ -12,7 +16,7 @@ public class MarketPostSimilarResponse {
     private Long postId;
     private String type;
     private String title;
-    private Integer price;
+    private BigDecimal price;
     private String thumbnailUrl;
     private LocalDateTime createdAt;
     private int similarityScore;
